requirements.txt
astropy==6.0.1
astropy-iers-data==0.2024.4.15.2.45.49
blinker==1.7.0
click==8.1.7
contourpy==1.2.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
DateTime==5.4
Deprecated==1.2.14
Flask==3.0.0
Flask-Limiter==3.5.0
fonttools==4.51.0
importlib-resources==6.1.1
itsdangerous==2.1.2
jdcal==1.4.1
Jinja2==3.1.2
kiwisolver==1.4.5
limits==3.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
mdurl==0.1.2
numpy==1.26.4
ordered-set==4.1.0
packaging==23.2
pillow==10.3.0
psutil==5.9.7
pyerfa==2.0.1.4
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
rich==13.7.0
scipy==1.13.0
setuptools==68.2.2
six==1.16.0
ttkbootstrap==1.10.1
typing_extensions==4.9.0
Werkzeug==3.0.1
wheel==0.41.2
wrapt==1.16.0
zope.interface==6.1
